Move crypto functions to NGPUtil
