+++ lib/drivers/medtronic600/NGPUtil.js
@@ -17,6 +17,7 @@
 
 import _ from 'lodash';
 import sundial from 'sundial';
+import crypto from 'crypto';
 
 // NGP = Next Generation Pump. It's the acronym that Medtronic use in their code for the 600-series.
 export class NGPTimestamp {
@@ -271,7 +272,7 @@ export class NGPConstants {
   }
 }
 
-export class NGPLinkKeyCipher {
+export class NGPLinkCipher {
   static randomBuffer(length) {
     return Buffer.from(Array.from({
       length,
@@ -283,7 +284,7 @@ export class NGPLinkKeyCipher {
       throw new TypeError('Requested cipher length must be 55 bytes or larger');
     }
 
-    const ciphertext = NGPLinkKeyCipher.randomBuffer(cipherLength);
+    const ciphertext = NGPLinkCipher.randomBuffer(cipherLength);
 
     /* eslint-disable no-bitwise, no-plusplus, no-restricted-syntax */
     let pos = key.slice(-1) & 7;
@@ -324,6 +325,20 @@ export class NGPLinkKeyCipher {
     /* eslint-enable no-bitwise, no-restricted-syntax */
 
     return plaintext;
+  }
+
+  static encrypt(key, iv, clear) {
+    const cipher = crypto.createCipheriv('aes-128-cfb', key, iv);
+    let encrypted = cipher.update(clear, 'binary', 'hex');
+    encrypted += cipher.final('hex');
+    return Buffer.from(encrypted, 'hex');
+  }
+
+  static decrypt(key, iv, encrypted) {
+    const decipher = crypto.createDecipheriv('aes-128-cfb', key, iv);
+    let clear = decipher.update(encrypted, 'binary', 'hex');
+    clear += decipher.final('hex');
+    return Buffer.from(clear, 'hex');
   }
 }
