+++ lib/drivers/medtronic600/medtronic600Driver.js
@@ -276,6 +276,8 @@ class MinimedPumpSession {
     return digest;
   }
 
+  // TODO: (for a refactor), should the sequence incrementing
+  // be done as an event listener on an EventEmitter in a send() command?
   incEnvelopeSequenceNumber() {
     this.envelopeSequenceNumber += 1;
     if (this.envelopeSequenceNumber >= 255) {
@@ -598,7 +600,7 @@ class RequestLinkKeyResponse extends MinimedResponse {
   }
 
   linkKey(cnlModelAndSerial) {
-    return NGPUtil.NGPLinkKeyCipher.unpackLinkKey(this.packedLinkKey, cnlModelAndSerial);
+    return NGPUtil.NGPLinkCipher.unpackLinkKey(this.packedLinkKey, cnlModelAndSerial);
   }
 }
 
@@ -711,7 +713,7 @@ class TransmitPacketResponse extends NGPResponse {
     const encryptedPayloadSize = this.payload[0x38];
     const encryptedPayload = Buffer.from(this.payload.slice(0x39, 0x39 + encryptedPayloadSize));
     const decryptedPayload =
-      TransmitPacketResponse.decrypt(pumpSession.key, pumpSession.iv, encryptedPayload);
+      NGPUtil.NGPLinkCipher.decrypt(pumpSession.key, pumpSession.iv, encryptedPayload);
     debug('### DECRYPTED PAYLOAD', decryptedPayload.toString('hex'));
 
     // Check the decrypted payload's checksums
@@ -724,13 +726,6 @@ class TransmitPacketResponse extends NGPResponse {
     }
 
     this.decryptedPayload = decryptedPayload;
-  }
-
-  static decrypt(key, iv, encrypted) {
-    const decipher = crypto.createDecipheriv('aes-128-cfb', key, iv);
-    let clear = decipher.update(encrypted, 'binary', 'hex');
-    clear += decipher.final('hex');
-    return Buffer.from(clear, 'hex');
   }
 
   get comDCommand() {
@@ -916,6 +911,9 @@ class TransmitPacketRequest extends NGPRequest {
     };
   }
 
+  // TODO: (for a refactor) - optional config should be via fluent functions, and then we
+  // build. So, somthing like this:
+  // new TransmitPacketRequest(pumpSession, comDCommand).setHighSpeed(false).buildPacket(parameters)
   constructor(
     pumpSession, comDCommand, parameters, responseType = TransmitPacketResponse,
     highSpeedMode = true,
@@ -961,7 +959,7 @@ class TransmitPacketRequest extends NGPRequest {
 
     // Encrypt the ComD message
     const encryptedBuffer =
-      TransmitPacketRequest.encrypt(pumpSession.key, pumpSession.iv, transmitBuffer);
+      NGPUtil.NGPLinkCipher.encrypt(pumpSession.key, pumpSession.iv, transmitBuffer);
 
     const payloadBuffer = Buffer.concat(
       [envelopeBuffer, encryptedBuffer],
@@ -971,13 +969,6 @@ class TransmitPacketRequest extends NGPRequest {
     super(NGPMessage.COMMAND_TYPE.TRANSMIT_PACKET, pumpSession, payloadBuffer, responseType);
     this.setMaxRetries(TransmitPacketRequest.RETRIES);
     this.multipacketSession = null;
-  }
-
-  static encrypt(key, iv, clear) {
-    const cipher = crypto.createCipheriv('aes-128-cfb', key, iv);
-    let encrypted = cipher.update(clear, 'binary', 'hex');
-    encrypted += cipher.final('hex');
-    return Buffer.from(encrypted, 'hex');
   }
 
   async read80Message(hidDevice, readTimeout) {
@@ -1441,9 +1432,8 @@ class DeviceCharacteristicsResponse extends TransmitPacketResponse {
 }
 
 class DeviceCharacteristicsCommand extends TransmitPacketRequest {
-  // FIXME: Can I get rid of this after AdHoc works?
-  // Specify an optional responseType, since DeviceCharacteristicsCommand is
-  // also used for ad hoc pairing (See AdHocPairingCommand)
+  // DeviceCharacteristicsCommand can be used both in highSpeedMode (normal comms)
+  // and during ad hoc pairing (low speed)
   constructor(pumpSession, responseType = DeviceCharacteristicsResponse, highSpeedMode = true) {
     const params = Buffer.alloc(9);
     params[0] = 0x02;
@@ -1818,7 +1808,7 @@ class AdHocPairingCommand extends TransmitPacketRequest {
         {
           const tempLinkKey = response.decryptedPayload.slice(0x3, 0x13).reverse();
           debug(`Temp link key: ${tempLinkKey.toString('hex')}`);
-          const tempPackedLinkKey = NGPUtil.NGPLinkKeyCipher
+          const tempPackedLinkKey = NGPUtil.NGPLinkCipher
             .packLinkKey(tempLinkKey, this.pumpSession.bcnlSerialNumber, 55);
           debug(`Temp packed link key: ${tempPackedLinkKey.toString('hex')}`);
           await Timer.wait(TransmitPacketRequest.SEND_DELAY_MS);
